Odstranenie vacsiny warning-ov z kodu a zakomentoval som toast ktory otravuje ze sa ma pripojit k zariadeniu, musim mu najst lepsie miesto

+++ app/src/main/java/com/example/smartfit/screens/ActivityDetailScreen.kt
@@ -171,7 +171,7 @@ fun ActivityDetailScreen(
                             )
                         )
                         Heading1(
-                            LocalTime.ofSecondOfDay((training.trainingDuration / 1000).toLong())
+                            LocalTime.ofSecondOfDay((training.trainingDuration / 1000))
                                 .format(DateTimeFormatter.ofPattern("HH:mm:ss"))
                         )
                     }
